Can we improve the descriptor handling?
Descriptors are used in

* the navigation,
* the result views,
* reference and identifiy things in data exchange

For the first two points, we want to show the most important
information of a data set. For the last point, we want to
identifiy a dataset in both directions: import & export

## id, refId, lastChange, version, name, description, modelType
All these fields should stay in the descriptors but we may can
simplify the query protocoll a bit here. Also, performance tests
would be good for the queries...

## Categories
The navigation and filtering by category is important for openLCA.
This is why storing the category ID in the descriptor probably is
a good idea. But for the display we often need to show (a part) of
the category path. Instead of building this path always dynamically
we may should store it in a field `categoryPath::String[]` of the
descriptor.

## Other fields
We should store them directly as strings not via IDs; e.g. flow units,
location codes --> we may can then drop the EntityCache? This would
probably simplify the API in a lot of places

## The descriptor protocoll in the DAOs

### RootEntityDao
* getDescriptors()
* getDescriptor(long id)
* getDescriptors(Set<Long> ids)
* getDescriptorForRefId(String refId)
* getDescriptorsForRefIds(Set<String> refIds)

### CategorizedEntityDao
* getDescriptors(Optional<Category> category)

=> more?